requirements.txt
langchain ==0.2.1
langchain_community == 0.2.1
langchain_core == 0.2.2
chromadb == 0.4.22
sentence-transformers
bs4 == 0.0.2
streamlit==1.35.0
python-dotenv==1.0.1
#pysqlite3-binary == 0.5.2
langchain_huggingface == 0.0.1